fixed small bug in pawn generation, perft testing works at around 155 million nodes per second

+++ src/chess/board/state.zig
@@ -295,7 +295,7 @@ pub const State = struct {
     hash: u64,
     castling: CastlingRights,
     enpassant: Square,
-    half_moves: u8,
+    half_moves: usize,
     captured_piece: Piece,
 };
